Add conversation sandbox adapter
